requirements.txt
jsonschema == 4.17.3
ops >= 2.4.1
pydantic == 1.10.10
ops-lib-pgsql >= 1.4
psycopg2-binary == 2.9.6
requests>=2,<3
